+++ Gruntfile.js
@@ -41,8 +41,8 @@ module.exports = function (grunt) {
             }
         },
 
-        copy:{
-            iconfont:{
+        copy: {
+            iconfont: {
                 expand: true,
                 cwd: 'app/styles/fonts',
                 dest: 'dist/styles/fonts',
@@ -56,9 +56,40 @@ module.exports = function (grunt) {
             port: 8889,
             testport: 9091,
             livereload: 9999
+        },
+
+        replace: {
+            development: {
+                options: {
+                    patterns: [{
+                        json: grunt.file.readJSON('config/environments/development.json')
+                    }]
+                },
+                files: [{
+                    expand: true,
+                    flatten: true,
+                    src: ['config/envConfig.js'],
+                    dest: 'app/scripts/config/'
+                }]
+            },
+            production: {
+                options: {
+                    patterns: [{
+                        json: grunt.file.readJSON('config/environments/production.json')
+                    }]
+                },
+                files: [{
+                    expand: true,
+                    flatten: true,
+                    src: ['config/envConfig.js'],
+                    dest: 'app/scripts/config/'
+                }]
+            }
         }
     });
 
-    grunt.registerTask('test', ['devbliss-karma:jenkins']);
-    grunt.registerTask('buildpullr', ['build', 'copy:iconfont']);
+    grunt.registerTask('test', ['replace:development', 'devbliss-karma:jenkins']);
+    grunt.registerTask('buildpullr', ['replace:production', 'build', 'copy:iconfont']);
+    grunt.task.renameTask('serve', 'devbliss-serve');
+    grunt.registerTask('serve', ['build', 'replace:development', 'devbliss-serve']);
 };