+++ app/app_components/translateModule/services/translateConst_tr.js
@@ -31,9 +31,10 @@ angular.module('translateModule')
             headline: 'Giriş',
             errorMessage: 'Giriş yapılamadı',
             username: 'Kullanıcı adı',
-            btn:{
+            btn: {
                 login: 'Giriş yap'
-            }
+            },
+            oauthGithub: 'Login with GitHub'
         },
         dashboard: {
             headline: {
@@ -79,7 +80,7 @@ angular.module('translateModule')
             }
         },
         settings: {
-            headline:{
+            headline: {
                 repo: 'Depo filtresi',
                 settings: 'Ayarlar'
             },