+++ test/spec/app_components/loginModule/services/userServiceSpec.js
@@ -75,56 +75,9 @@ describe('userService', function () {
         });
     });
 
-    describe('getUsersForLogin', function () {
-        var expectedUrl = '/api/users',
-            successPayload = {
-                data: [
-                    {id: 12345, username: 'testUser', avatarUrl: 'http://www.jira.de'},
-                    {id: 42, username: 'user2', avatarUrl: 'http://example.org'}
-                ],
-                status: 200
-            },
-            errorPayload = {
-                data: {errorKey: 'AnyErrorK', errorMessage: 'no users available'},
-                status: 400
-            };
-
-        beforeEach(function () {
-            response = $httpBackend.expectGET(expectedUrl).respond(successPayload.status, successPayload.data);
-        });
-
-        it('calls correct URL', function () {
-            service.getUsersForLogin();
-            $httpBackend.flush();
-        });
-
-        it('returns correct data', function () {
-            var result = null;
-
-            service.getUsersForLogin().then(function (users) {
-                result = users;
-            });
-
-            $httpBackend.flush();
-            expect(result).toEqual(successPayload.data);
-        });
-
-        it('forwards error', function () {
-            response.respond(errorPayload.status, errorPayload.data);
-
-            service.getUsersForLogin().then(function (successResponse) {
-                expect(successResponse).toBeNull();
-            }, function (errorResponse) {
-                expect(errorResponse.data).toEqual(errorPayload.data);
-            });
-
-            $httpBackend.flush();
-        });
-    });
-
-    describe('logInUser', function () {
-        var user = {id: 12345, username: 'testUser', avatarUrl: 'http://www.jira.de'},
-            expectedUrl = '/api/users/login/' + user.id,
+    describe('authenticateWithGithubAndLogInUser', function() {
+        var githubCode = 'some-random-code-123',
+            expectedUrl = '/api/users/oauth/github/' + githubCode,
             successPayload = {
                 data: {
                     id: 42, username: 'user2', avatarUrl: 'http://example.org', userSettingsDto: {
@@ -143,17 +96,14 @@ describe('userService', function () {
         });
 
         it('calls correct URL', function () {
-            $httpBackend.expectGET('/api/users/me').respond(successPayload.status, successPayload.data);
-
-            service.logInUser(user);
+            service.authenticateWithGithubAndLogInUser(githubCode);
             $httpBackend.flush();
         });
 
         it('returns correct data', function () {
-            $httpBackend.expectGET('/api/users/me').respond(successPayload.status, successPayload.data);
             var success = null;
 
-            service.logInUser(user).then(function () {
+            service.authenticateWithGithubAndLogInUser(githubCode).then(function () {
                 success = true;
             });
 
@@ -164,7 +114,7 @@ describe('userService', function () {
         it('forwards error', function () {
             response.respond(errorPayload.status, errorPayload.data);
 
-            service.logInUser(user).then(function (successResponse) {
+            service.authenticateWithGithubAndLogInUser(githubCode).then(function (successResponse) {
                 expect(successResponse).toBeNull();
             }, function (errorResponse) {
                 expect(errorResponse.data).toEqual(errorPayload.data);